fix: move error.ts to not be deleted by Makefile
