+++ gnoboard/src/grpc/error.ts
