+++ gnoboard/src/screens/switch-accounts/ReenterPassword.tsx
@@ -1,4 +1,4 @@
-import { GRPCError } from '@gno/api/error';
+import { GRPCError } from '@gno/grpc/error';
 import { ErrCode } from '@gno/api/rpc_pb';
 import Alert from '@gno/components/alert';
 import Button from '@gno/components/buttons';
